Added more functionalities to SimulationParameter base classes
