+++ SimulationWrapper.py
@@ -1,7 +1,6 @@
 import b2luigi
 import os
-from typing import Dict
-from utils.htcondor_settings import local_condor_settings, global_htcondor_settings
+import numpy
 
 
 class StartSimulationTask(b2luigi.Task):
@@ -20,33 +19,6 @@ class StartSimulationTask(b2luigi.Task):
 
 
 class SimulationWrapperTask(b2luigi.WrapperTask):
-    batch_system = "htcondor"
-    _resources 
-
-    @property
-    def htcondor_settings(self) -> Dict[str, str]:
-        base_settings = global_htcondor_settings
-        base_settings.update({
-            f"request_{resource}": str(value) for resource, value in self._resources.items() 
-        })
-        base_settings.update({
-            "transfer_executable": "true",
-            "transfer_input_files": "",
-            "when_to_transfer_output": "ON_SUCCESS",
-            "transfer_output_files": "remote_setup.sh",
-            "JobBatchName": self.stage,
-        })
-        return base_settings
-    
-    @property
-    def resources(self) -> Dict[str, int]:
-        if self.batch_system == "local":
-            res = {key: value for key, value in self._resources.items() if key in ["cpus", "memory", "gpus"]}
-            res.update({
-                "local_workers": 1})
-            return res
-        else:
-            return {"htcondor": 1}
 
     def requires(self):
         for i in range(3):
