+++ utils/SimulationHelpers.py
@@ -1,18 +1,52 @@
-from typing import Type
+from typing import Type, Dict, List
+import json
 
 
 class SimulationParameter():
+    """ Base class for all parameters used in the simulation
+
+    TODO Write warnings in case the base class is used directly in the dictionary
+    ref: https://stackoverflow.com/questions/46092104/subclass-in-type-hinting
+    """
 
     def __init__(self, starting_value):
         """ Initialize a new general parameter
         """
         self.starting_value = starting_value
 
+    def to_dict(self) -> Dict:
+        """ Convert to dictionary
+        """
+        return {
+            "starting_value": self.starting_value
+        }
+
+    @classmethod
+    def from_dict(cls, parameter_dict: Dict) -> Dict:
+        """ Create from dictionary
+        """
+        return cls(**parameter_dict)
+
 
 class SimulationParameterDictionary():
 
     def __init__(self):
-        """ Initialize an empty dictionary with no parameters
+        """ Initialize an empty list with no parameters
+        """
+        self.parameter_list: List[Type[SimulationParameter]] = []
+
+    def add_parameter(self, simulation_parameter: Type[SimulationParameter]):
+        """ Add a parameter to the dictionary 
+        """
+        self.parameter_list.append(simulation_parameter)
+
+    def to_dict(self) -> Dict:
+        """ Converts to dictionary
+        """
+        return {"Parameters": [parameter.to_dict() for parameter in self.parameter_list]}
+
+    def from_dict(self):
+        """ Create an instance from dictionary
         """
 
 
@@ -24,8 +58,6 @@ class SimulationIO():
 
     def write_json():
         """ Write the dictionary of parameters to a json file for safekeeping
-
-
         """
 
     def read_json():
